requirements.txt
python_magic==0.4.27
